services/server/requirements.txt
gunicorn==20.0.4
psycopg2-binary==2.8.5
Flask==1.1.2
Flask-Cors==3.0.8
flask-migrate==2.5.3
Flask-SQLAlchemy==2.4.1